Completed scrolling with camera bounds and map bounds features
A new major addition to the camera system! This took me an hour to
figure out, and was pretty annoying, but is really awesome. You can now
set bounds to the camera in which the player can move around before the
camera starts moving with the player. There is also a map bounds that
the camera cannot leave. For example, if you set the map bounds to the
walls of the level, the camera will not show anything outside of those
walls.